refactor(adapters): extract shared K8s quantity utils and fix dict aliasing
Extract duplicated parseK8sQuantity, formatHumanBytes, and formatResourceValue
into shared web/src/utils/k8sQuantity.ts. Fix dict aliasing bug in
ResourceContributor where cpu_mem_requests was shared by reference between
devrunner and skuld values. Update test_rest_resources to match GPU removal
from StaticResourceProvider.
