+++ web/src/utils/index.ts
@@ -12,3 +12,9 @@ export {
 } from './formatters';
 export { serializePresetYaml, parsePresetYaml } from './presetYaml';
 export type { PresetRuntimeFields } from './presetYaml';
+export {
+  BYTE_MULTIPLIERS,
+  parseK8sQuantity,
+  formatHumanBytes,
+  formatResourceValue,
+} from './k8sQuantity';
